refactor: merged get by userid
